Better error reporting for Store classes that don't meet constraints (#278)
* refactored tests to be file driven rather than putting as strings in test files

* fixing dart analyzer issues

* upgrading test_coverage

* invalid @action checks

* adding more error reporting
* fixing RegExp to ensure only the starting '_' is replaced
