+++ mobx_codegen/test/data/invalid_annotated_store.dart
@@ -1,0 +1,9 @@
+library generator_sample;
+
+import 'package:mobx/mobx.dart';
+
+part 'generator_sample.g.dart';
+
+@store
+// ignore: unused_element
+class PublicUser {}
