+++ mobx_codegen/test/data/invalid_annotated_store_multiple.dart
@@ -1,0 +1,13 @@
+library generator_sample;
+
+import 'package:mobx/mobx.dart';
+
+part 'generator_sample.g.dart';
+
+@store
+// ignore: unused_element
+class PublicUser1 {}
+
+@store
+// ignore: unused_element
+class PublicUser2 {}
