+++ mobx_codegen/test/data/invalid_annotated_store_output.txt
@@ -1,0 +1,2 @@
+Could not make class "PublicUser" observable. Changes needed:
+  1. Classes that use the @store annotation must be marked private. Affected classes: "PublicUser".
