+++ mobx_codegen/test/data/invalid_mixin_store.dart
@@ -1,0 +1,10 @@
+library generator_sample;
+
+import 'package:mobx/mobx.dart';
+
+part 'generator_sample.g.dart';
+
+class NonAbstractUser = NonAbstractUserBase with _$NonAbstractUserBase;
+
+// ignore: unused_element
+class NonAbstractUserBase with Store {}
