+++ mobx_codegen/test/data/invalid_mixin_store_output.txt
@@ -1,0 +1,2 @@
+Could not make class "NonAbstractUser" observable. Changes needed:
+  1. Classes that use the Store mixin must be marked abstract. Affected classes: "NonAbstractUserBase".
