+++ mobx_codegen/test/invalid_annotations_test.dart
@@ -10,6 +10,20 @@ void main() {
           source: './data/invalid_store_declaration.dart',
           output: './data/invalid_store_declaration_output.txt'),
       const TestInfo(
+          description: 'Errors when a non-abstract class mixes in a Store',
+          source: './data/invalid_mixin_store.dart',
+          output: './data/invalid_mixin_store_output.txt'),
+      const TestInfo(
+          description:
+              'Errors when there is an @store annotation on a public class',
+          source: './data/invalid_annotated_store.dart',
+          output: './data/invalid_annotated_store_output.txt'),
+      const TestInfo(
+          description:
+              'Errors when there are multiple @store annotations on public classes',
+          source: './data/invalid_annotated_store_multiple.dart',
+          output: './data/invalid_annotated_store_multiple_output.txt'),
+      const TestInfo(
           description:
               'Errors when there is a single invalid @computed annotation',
           source: './data/invalid_computed_single.dart',
