+++ mobx_codegen/test/mobx_codegen_test.dart
@@ -18,17 +18,6 @@ void main() {
       expect(await generate(source), isEmpty);
     });
 
-    test('ignores public classes annotated with @store', () async {
-      const source = """
-        @store
-        class MyClass {
-          void foobar() => 'Hello';
-        }
-      """;
-
-      expect(await generate(source), isEmpty);
-    });
-
     test('ignores when there is no class other than the abstract Store',
         () async {
       final source = await readFile('./data/only_abstract_store.dart');
@@ -58,7 +47,8 @@ void main() {
           source: './data/valid_generic_annotated_store_input.dart',
           output: './data/valid_generic_annotated_store_output.dart'),
       const TestInfo(
-          description: 'generates correct types for a @store referencing another @store',
+          description:
+              'generates correct types for a @store referencing another @store',
           source: './data/valid_annotated_store_referencing_store_input.dart',
           output: './data/valid_annotated_store_referencing_store_output.dart'),
       const TestInfo(
